requirements.txt
# --- Web Framework & Production Server ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0              # Essential for process management on Render
python-multipart==0.0.9       # Required for file uploads in ingestion_service

# --- Computation & Analytical Engines (Execution Layer) ---
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0                # Vectorized dataframe engine
scipy>=1.12.0                 # Mathematical precision & anomaly detection
duckdb>=0.10.0                # In-process analytical engine
pyarrow>=14.0.0               # Parquet support for DuckDB/Polars
scikit-learn>=1.4.0           # ML utilities

# --- Intelligence & RAG (Intelligence Layer) ---
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
sentence-transformers>=2.5.0
tiktoken>=0.6.0
chromadb>=0.4.24
faiss-cpu>=1.7.4

# --- Database & Multi-Tenant Infrastructure ---
supabase>=2.4.0               # Auth and DB management
postgrest>=0.10.0
psycopg2-binary>=2.9.9        # Postgres adapter
sqlalchemy>=2.0.27            # ORM for models.py
pgvector>=0.2.5               # FIX: Required for Vector type in models.py
pydantic>=2.6.0
pydantic-settings>=2.2.0

# --- Storage & Cloud Strategy (Modular Infrastructure) ---
boto3==1.34.51                # FIX: Pinned to mathematically match s3fs 2024.2.0
s3fs==2024.2.0                
fsspec==2024.2.0         

# --- Security & Utilities ---
python-dotenv==1.0.1
bcrypt==4.1.2
pyjwt==2.8.0
cryptography==42.0.5
python-dateutil==2.8.2
croniter==2.0.1               # Schedule handling
openpyxl==3.1.2               # Excel ingestion support
httpx==0.27.0                 # Async HTTP client for AI streaming/Supabase calls

# --- Observability & Telemetry ---
prometheus-client==0.20.0
structlog==24.1.0
snowflake-connector-python[pandas]==3.0.3
loguru==0.6.0                 # Advanced logging capabilities
sentry-sdk==1.28.0            # Error tracking and performance monitoring